Use fake objects in tests
