+++ data/src/test/kotlin/io/github/fobo66/data/repositories/RatingRepositoryImplTest.kt
@@ -1,8 +1,7 @@
 package io.github.fobo66.data.repositories
 
-import io.github.fobo66.data.entities.MatchmakingRating
-import io.github.fobo66.data.source.PersistenceDataSource
-import io.github.fobo66.data.source.PreferenceDataSource
+import io.github.fobo66.data.fake.FakePersistenceDataSource
+import io.github.fobo66.data.fake.FakePreferenceDataSource
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runTest
 import kotlin.test.BeforeTest
@@ -11,29 +10,9 @@ import kotlin.test.assertNull
 
 class RatingRepositoryImplTest {
 
-    private val persistenceDataSource = object : PersistenceDataSource {
-        override suspend fun loadRating(id: Long): MatchmakingRating? {
-            return null
-        }
+    private val persistenceDataSource = FakePersistenceDataSource()
 
-        override suspend fun saveRating(rating: MatchmakingRating): Long {
-            return 0L
-        }
-    }
-
-    private val preferenceDataSource = object : PreferenceDataSource {
-        override suspend fun getBoolean(key: String, defaultValue: Boolean): Boolean {
-            return defaultValue
-        }
-
-        override suspend fun saveBoolean(key: String, value: Boolean) = Unit
-
-        override suspend fun getLong(key: String, defaultValue: Long): Long {
-            return defaultValue
-        }
-
-        override suspend fun saveLong(key: String, value: Long) = Unit
-    }
+    private val preferenceDataSource = FakePreferenceDataSource()
 
     private lateinit var ratingRepository: RatingRepository
